LayoutTests/editing/inserting/font-size-clears-from-typing-style-expected.txt

Tests that font-size in typingStyle is correctly cleared. See https://bugs.webkit.org/show_bug.cgi?id=26279.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 1 is 1
PASS 3 is 3
PASS successfullyParsed is true

TEST COMPLETE
BA
